requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
braceexpand==0.1.7
cachetools==5.5.0
click==8.1.7
cloudpickle==3.0.0
cython==3.0.11
decorator==4.4.2
faster-fifo==1.4.7
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.34.0
google-auth-oauthlib==0.4.6
grpcio==1.65.5
gym==0.23.0
gym-notices==0.0.8
habitat-sim==0.3.1
hydra-core==1.3.2
ifcfg==0.24
importlib-metadata==8.3.0
iniconfig==2.0.0
jinja2==3.1.4
lmdb==1.5.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
objectio==0.2.29
omegaconf==2.3.0
opencv-python==4.10.0.84
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.2.0
protobuf==3.20.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pybullet==3.0.4
pygame==2.0.1
pytest==8.3.2
pyyaml==6.0.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.2.0
shellingham==1.5.4
simplejson==3.19.3
sympy==1.13.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
typer==0.12.4
webdataset==0.1.40
websockets==13.1
werkzeug==3.0.3
wget==3.2
yarl==1.14.0
transformers==4.46.0